Use a backgroundjob to move avatars
