+++ lib/private/Repair/NC11/MoveAvatars.php
@@ -23,26 +23,14 @@
 namespace OC\Repair\NC11;
 
 use OC\SystemConfig;
-use OCP\Files\File;
-use OCP\Files\Folder;
-use OCP\Files\IAppData;
-use OCP\Files\IRootFolder;
-use OCP\Files\NotFoundException;
-use OCP\IUser;
-use OCP\IUserManager;
+use OCP\BackgroundJob\IJobList;
 use OCP\Migration\IOutput;
 use OCP\Migration\IRepairStep;
 
 class MoveAvatars implements IRepairStep {
 
-	/** @var IUserManager */
-	private $userManager;
-
-	/** @var IRootFolder */
-	private $rootFolder;
-
-	/** @var IAppData */
-	private $appData;
+	/** @var IJobList */
+	private $jobList;
 
 	/** @var SystemConfig */
 	private $systemConfig;
@@ -50,18 +38,12 @@ class MoveAvatars implements IRepairStep {
 	/**
 	 * MoveAvatars constructor.
 	 *
-	 * @param IUserManager $userManager
-	 * @param IRootFolder $rootFolder
-	 * @param IAppData $appData
+	 * @param IJobList $jobList
 	 * @param SystemConfig $systemConfig
 	 */
-	public function __construct(IUserManager $userManager,
-								IRootFolder $rootFolder,
-								IAppData $appData,
+	public function __construct(IJobList $jobList,
 								SystemConfig $systemConfig) {
-		$this->userManager = $userManager;
-		$this->rootFolder = $rootFolder;
-		$this->appData = $appData;
+		$this->jobList = $jobList;
 		$this->systemConfig = $systemConfig;
 	}
 
@@ -69,66 +51,14 @@ class MoveAvatars implements IRepairStep {
 	 * @return string
 	 */
 	public function getName() {
-		return 'Move avatars to AppData folder';
+		return 'Add mover avatar background job';
 	}
 
 	public function run(IOutput $output) {
 		if ($this->systemConfig->getValue('enable_avatars', true) === false) {
 			$output->info('Avatars are disabled');
 		} else {
-			$output->startProgress($this->userCount());
-			$this->moveAvatar($output);
-			$output->finishProgress();
+			$this->jobList->add(MoveAvatarsBackgroundJob::class);
 		}
-	}
-
-	private function moveAvatar(IOutput $output) {
-		$this->userManager->callForAllUsers(function (IUser $user) use ($output) {
-			if ($user->getLastLogin() !== 0) {
-				$uid = $user->getUID();
-
-				\OC\Files\Filesystem::initMountPoints($uid);
-				/** @var Folder $userFolder */
-				$userFolder = $this->rootFolder->get($uid);
-
-				try {
-					$userData = $this->appData->getFolder($uid);
-				} catch (NotFoundException $e) {
-					$userData = $this->appData->newFolder($uid);
-				}
-
-
-				$regex = '/^avatar\.([0-9]+\.)?(jpg|png)$/';
-				$avatars = $userFolder->getDirectoryListing();
-
-				foreach ($avatars as $avatar) {
-					/** @var File $avatar */
-					if (preg_match($regex, $avatar->getName())) {
-						/*
-						 * This is not the most effective but it is the most abstract way
-						 * to handle this. Avatars should be small anyways.
-						 */
-						$newAvatar = $userData->newFile($avatar->getName());
-						$newAvatar->putContent($avatar->getContent());
-						$avatar->delete();
-					}
-				}
-			}
-			$output->advance();
-		});
-	}
-
-	/**
-	 * @return int
-	 */
-	private function userCount() {
-		$backends = $this->userManager->countUsers();
-		$count = 0;
-
-		foreach ($backends as $backend => $amount) {
-			$count += $amount;
-		}
-
-		return $count;
 	}
 }
